Create new core.magics package and start populating with history.
